Ajout des contrats dans l'affichage d'un seul utilisateur avec désactivation du crud contrat

+++ PHP/VIEW/LISTE/ListeContrats.php
@@ -1,5 +1,6 @@
 <?php
 
+$formtype = ($_GET['page'] == 'FormUtilisateurs' && $_GET['mode'] == 'Afficher' ? true : false);
 
 echo '<div><section class="colonne">';
 
@@ -7,13 +8,13 @@ $objets = ContratsManager::getList(null, ['idUtilisateur' => $_GET['id']], null,
 echo '<div class="noDisplay NbEltParPage">' . Parametres::getNbEltParPage() . '</div>';
 echo '<div class="bigEspace"></div>';
 echo '<div class="bigEspace"></div>';//Création du template de la grid
-echo '<div class="grid-col-7 gridListe">';
+echo '<div class="grid-col-'.($formtype?'4-nobutton':7).' gridListe">';
 
-echo '<div class="caseListe titreListe grid-columns-span-7">'.texte("Liste des Contrats").'</div>';
-echo '<div class="caseListe grid-columns-span-7">
+echo '<div class="caseListe titreListe grid-columns-span-'.($formtype?'4-nobutton':7).'">'.texte("Liste des Contrats").'</div>';
+echo '<div class="caseListe grid-columns-span-'.($formtype?'4-nobutton':7).'">
 <div></div>
 <div class="bigEspace"></div>
-<div class="caseListe"><a href="index.php?page=FormContrats&mode=Ajouter&idutilisateur='.$_GET['id'].'"><i class="fas fa-plus"></i></a></div>
+<div class="caseListe">'.(!$formtype?'<a href="index.php?page=FormContrats&mode=Ajouter&idutilisateur='.$_GET['id'].'"><i class="fas fa-plus"></i></a>':'').'</div>
 <div></div>
 </div>';
 
@@ -25,7 +26,7 @@ echo '<div class="caseListe labelListe" data-name= "DateFinContrat">DateFinContr
 //Remplissage de div vide pour la structure de la grid
 echo '<div class="caseListe"></div>';
 echo '<div class=" caseListe texteClair "></div><div class="mini" id="nbEnregs"></div> ';
-echo '</div><div class="grid-col-7 gridListe grid-contenu">';
+echo '</div><div class="grid-col-'.($formtype?'4-nobutton':7).' gridListe grid-contenu">';
 
 // Affichage des enregistrements de la base de données
 foreach ($objets AS $value)
@@ -34,20 +35,22 @@ foreach ($objets AS $value)
 	echo '<div class="donnees ">'.UtilisateursManager::findById($value->getIdUtilisateur())->getNomUtilisateur().'</div>';
 	echo '<div class="donnees ">'.$value->getDateDebutContrat().'</div>';
 	echo '<div class="donnees ">'.$value->getDateFinContrat().'</div>';
-	echo '<a href="index.php?page=FormContrats&mode=Afficher&id='.$value->getIdContrat().'&idutilisateur='.$_GET['id'].'"><i class="fas fa-file-contract"></i></a>';
-										
-	echo '<a href="index.php?page=FormContrats&mode=Modifier&id='.$value->getIdContrat().'&idutilisateur='.$_GET['id'].'"><i class="fas fa-pen"></i></a>';
-										
-	echo '<a href="index.php?page=FormContrats&mode=Supprimer&id='.$value->getIdContrat().'&idutilisateur='.$_GET['id'].'"><i class="fas fa-trash-alt"></i></a>';
+	if (!$formtype){
+		echo '<a href="index.php?page=FormContrats&mode=Afficher&id='.$value->getIdContrat().'&idutilisateur='.$_GET['id'].'"><i class="fas fa-file-contract"></i></a>';
+
+		echo '<a href="index.php?page=FormContrats&mode=Modifier&id='.$value->getIdContrat().'&idutilisateur='.$_GET['id'].'"><i class="fas fa-pen"></i></a>';
+
+		echo '<a href="index.php?page=FormContrats&mode=Supprimer&id='.$value->getIdContrat().'&idutilisateur='.$_GET['id'].'"><i class="fas fa-trash-alt"></i></a>';
+	}
 }
 echo '</div></div>';
 
-//Derniere ligne du tableau (bouton retour)
+//Dernière ligne du tableau (bouton retour)
 echo '<div class="bigEspace"></div>';
 
 echo '<div class="caseListe grid-columns-span-7">
 	<div></div>
-	<a href="index.php?page=Accueil"><button><i class="fas fa-house fa-rotate-180"></i></button></a>
+	<a href="index.php?page=ListeUtilisateurs"><button><i class="fas fa-house fa-rotate-180"></i></button></a>
 	<div></div>
 </div>';
 
@@ -56,4 +59,4 @@ echo '<div class="bigEspace grid-columns-span-9"></div>';
 echo '<div class="bigEspace grid-columns-span-9 pagination"></div>';
 
 echo '<div class="bigEspace grid-columns-span-9"></div>';
-echo'</div>'; 
+echo'</div>';
